Move AppBooks component to matters dir
